requirements.txt
xlrd
pytz
django_mysql==3.2.0
Whoosh==2.7.4
jieba==0.39
Pillow==6.1.0
django_haystack==2.8.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
PyMySQL==0.9.3
oauthlib==3.1.0
mysqlclient==1.4.4
Django==2.2.4